requirements.txt
google-cloud-bigquery
pandas
sqlalchemy
jupyter
matplotlib
seaborn
db-dtypes
